Create Activity_4.2.2
TASK-
Complete the code to find if a given number is a prime number? The program will take a positive integer greater than 1 as input and indicate if it is a prime number by saying "prime", and if it is not a prime number saying "not a prime". Note there are 3 places in the given code that you need to fix for this code to work properly and give the expected output. 

GIVEN-
i = int(input())
j = 5 # fix the code (1) 
while (j <= (i/j)):
    if not(i%j): 
        print("not a prime")
        continue # fix the code (2)
    j = j + 2 # fix the code (3)
if (j > i/j): 
    print ("prime")